backend/requirements.txt
amqp==5.1.1
asgiref==3.4.1
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
bleach==5.0.1
boto3==1.24.39
botocore==1.27.96
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==3.2.9
django-3-jet==1.0.8
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==3.11.0
django-environ==0.4.5
django-extensions==3.2.0
django-filter==21.1
git+https://github.com/ecometrica/django-hashedfilenamestorage@master
django-js-asset==2.0.0
django-material-admin==1.8.6
django-redis==5.2.0
django-storages==1.13.2
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.11
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==20.1.0
idna==3.4
ipython==8.3.0
jedi==0.18.2
jmespath==1.0.1
kombu==5.2.4
matplotlib-inline==0.1.6
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.3.3
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
traitlets==5.8.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
